\name{generatetonetworkfromvel}
\alias{generatetonetworkfromvel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generates a time-ordered network from a data frame listing all directed edges. An internal function.
}
\description{
-
}
\usage{
generatetonetworkfromvel(vel)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vel}{
A data frame listing all directed edges
}
}



\author{Benjamin Blonder \email{bblonder@email.arizona.edu}.}


%% ~Make other sections like Warning with \section{Warning }{....} ~


\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
