\name{generatevertexedgelist}
\alias{generatevertexedgelist}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generates a data frame listing all directed edges in a time-ordered network from an observed interaction list. An internal function.
}
\description{
-
}
\usage{
generatevertexedgelist(raw, allindivs)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{raw}{
A data frame of events
}
  \item{allindivs}{
A vector of names
}
}



\author{Benjamin Blonder \email{bblonder@email.arizona.edu}.}


%% ~Make other sections like Warning with \section{Warning }{....} ~



