\name{timeordered-package}
\alias{timeordered-package}
\alias{timeordered}
\docType{package}
\title{
Time-ordered and time-aggregated network analyses
}
\description{
Methods for incorporating time into network analysis. Construction of time-ordered networks (temporal graphs). Shortest-time and shortest-path-length analyses. Resource spread calculations. Data resampling and rarefaction for null model construction. Reduction to time-aggregated networks with variable window sizes; application of common descriptive statistics to these networks. Vector clock latencies. Plotting functionalities & full compatibility with all network methods in the igraph library.
}
\details{
\tabular{ll}{
Package: \tab timeordered\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2011-03-27\cr
License: GPL-3
LazyLoad: \tab yes\cr
}
}
\author{
Benjamin Blonder

Maintainer: Benjamin Blonder <bblonder@email.arizona.edu>
}
\references{
Blonder & Dornhaus. Time-ordered networks reveal limitations to information flow in ant colonies. PLoS One (2011) in press.
}
\keyword{ package }
\seealso{
\code{\link{igraph}}
}