\name{timeordered-package}
\alias{timeordered-package}
\alias{timeordered}
\docType{package}
\title{
Time-ordered and time-aggregated network analyses
}
\description{
Methods for incorporating time into network analysis. Construction of time-ordered networks (temporal graphs). Shortest-time and shortest-path-length analyses. Resource spread calculations. Data resampling and rarefaction for null model construction. Reduction to time-aggregated networks with variable window sizes; application of common descriptive statistics to these networks. Vector clock latencies. Plotting functionalities.
}
\details{
\tabular{ll}{
Package: \tab timeordered\cr
Type: \tab Package\cr
Version: \tab 0.9.7\cr
Date: \tab 2014-10-22\cr
License: GPL-3
LazyLoad: \tab yes\cr
}
}
\author{
Benjamin Blonder

Maintainer: Benjamin Blonder <bblonder@email.arizona.edu>
}
\references{
Blonder B, Dornhaus A (2011) Time-Ordered Networks Reveal Limitations to Information Flow in Ant Colonies. PLoS ONE, 6(5), e20298.

Blonder, B., Wey, T. W., Dornhaus, A., James, R. and Sih, A. (2012), Temporal dynamics and network analysis. Methods in Ecology and Evolution, 3, 958 - 972.
}
\keyword{ package }
\seealso{
\code{\link{igraph}}
}